- Changed ITestLogcatLogger.kt to TestLogcatLogger.kt
