+++ logcat/src/nativeTest/kotlin/logcat/platformTestLogger.kt
@@ -1,5 +1,5 @@
 package logcat
 
-actual fun platformTestLogger(isLoggable: (LogPriority) -> Boolean): ITestLogcatLogger {
+actual fun platformTestLogger(isLoggable: (LogPriority) -> Boolean): TestLogcatLogger {
   return NativeTestLogcatLogger(isLoggable = isLoggable)
 }
